+++ tests/Resources/UserResource.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Mblarsen\LaravelRepository\Tests\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class UserResource extends JsonResource
+{
+    public function toArray($request)
+    {
+        return parent::toArray($request);
+    }
+}