+++ tests/Resources/UserResourceCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Mblarsen\LaravelRepository\Tests\Resources;
+
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class UserResourceCollection extends ResourceCollection
+{
+    public function toArray($request)
+    {
+        return parent::toArray($request);
+    }
+}
+