+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to `laravel-repository` will be documented in this file
 
 ## next
 
+-   feat: support for JsonResource (see [README resources section](README.md#resources)
+
 ## 0.3.0
 
 -   feat: add [`list()`](README.md#list)